backend/pyproject.toml
[tool.poetry]
name = "job-search-automation-backend"
version = "0.1.0"
description = "Backend for Job Search Automation project"
authors = ["Chris Kuffo <ckuffo@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
supabase = "^2.9.0"
pydantic-settings = "^2.5.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.11.2"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
